+++ connectToAWSServerViaSSH.sh
@@ -1,1 +1,1 @@
-ssh -i ~/.ssh/linux-demo.pem ec2-user@3.134.83.141
+ssh -i ~/.ssh/linux-demo.pem ec2-user@3.134.86.104