+++ src/main/java/com/github/jeffmagina/restaurant/main/Main.java
@@ -1,13 +1,10 @@
 package com.github.jeffmagina.restaurant.main;
 
 import java.io.File;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
+import java.util.ArrayList;
 
 import com.github.jeffmagina.io.IO;
+import com.github.jeffmagina.io.OrderRepo;
 import com.github.jeffmagina.restaurant.cashier.Cashier;
 import com.github.jeffmagina.restaurant.customerorder.CustomerOrder;
 
@@ -24,24 +21,18 @@ public class Main {
 		//Cashier interactions with customer
 		cashier.interaction(custOrder);
 		
-		//Testing reading from AWS database
-		String url = "jdbc:postgresql://3.134.83.141:5432/jeffMagina";
-		String user = "jeffMagina";
-		String password = "jeffMagina";
+		//store into file and database
+		cashier.storeOrder();
+		
+		ArrayList<CustomerOrder> custOrders = new ArrayList<CustomerOrder>();
+		custOrders.add(custOrder);
+		
+		
+		
 
-		try (Connection conn = DriverManager.getConnection(url, user, password);) {
-			Statement stmt = conn.createStatement();
-			ResultSet rs = stmt.executeQuery("select * from orderhistory");
-			while (rs.next()) {
-				System.out.println(rs.getString("order_id"));
-				System.out.println(rs.getString("name"));
-				System.out.println(rs.getString("ordercost"));
-				System.out.println(rs.getString("paymentamount"));
-				System.out.println(rs.getString("changegiven"));
-			}
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
+		
+		System.out.println("done");
+		
 
 	}
 }