+++ src/main/java/com/github/jeffmagina/restaurant/orderhistory/OrderHistory.java
@@ -1,5 +1,6 @@
 package com.github.jeffmagina.restaurant.orderhistory;
 
+import com.github.jeffmagina.io.OrderRepo;
 import com.github.jeffmagina.restaurant.customerorder.CustomerOrder;
 import java.util.ArrayList;
 //import com.github.jeffmagina.restaurant.customerorder.CustomerOrder;
@@ -7,18 +8,23 @@ import java.util.ArrayList;
 public class OrderHistory extends CustomerOrder {
 
 	static ArrayList<CustomerOrder> allOrders = new ArrayList<CustomerOrder>();
+	OrderRepo orderRepo = new OrderRepo();
 
 	public OrderHistory() {
 	}
 
 	public void storeOrder(CustomerOrder custOrder) {
-		OrderHistory.allOrders.add(custOrder);
+		
+		allOrders.add(custOrder);
 		// store in database
+		orderRepo.insertAll(allOrders);
+		
 	}
 
 	public void displayAllOrders() {
+		allOrders = orderRepo.readAll();
     	for(int i = 0; i < allOrders.size(); i++) {
-    	System.out.println(OrderHistory.allOrders.get(i));
+    	System.out.println(allOrders.get(i));
     	}
     }
 }